+++ protostar/protostar_toml/core/protostar_toml_section.py
@@ -7,7 +7,7 @@ class ProtostarTOMLSection(ABC):
 
     @staticmethod
     @abstractmethod
-    def get_name() -> str:
+    def get_section_name() -> str:
         ...
 
     @abstractmethod
