+++ protostar/protostar_toml/core/protostar_toml_writer.py
@@ -3,7 +3,7 @@ from typing import List, OrderedDict
 
 import tomli_w
 
-from protostar.protostar_toml._protostar_toml_section import ProtostarTOMLSection
+from protostar.protostar_toml.core.protostar_toml_section import ProtostarTOMLSection
 from protostar.protostar_toml.protostar_config import ProtostarConfig
 from protostar.protostar_toml.protostar_contracts import ProtostarContracts
 from protostar.protostar_toml.protostar_project import ProtostarProject
@@ -27,7 +27,7 @@ class ProtostarTOMLWriter:
         ]
 
         for section in sections:
-            result[section.get_name()] = section.to_dict()
+            result[section.get_section_name()] = section.to_dict()
 
         with open(path, "wb") as protostar_toml_file:
             tomli_w.dump(result, protostar_toml_file)
